backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6

# Computer Vision
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0

# Deep Learning
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.206

# Firebase
firebase-admin==6.2.0

# WebSocket
websockets==12.0

# Utilities
aiofiles==23.2.1
requests==2.31.0